requirements.txt
requests
Pillow
static-ffmpeg